moved form one loop to 2 and added envs for that

+++ src/solscan/structs/market.rs
@@ -5,11 +5,11 @@ pub struct MarketPrice {
     #[serde(rename = "priceUsdt")]
     pub price_usdt: Option<f64>,
     #[serde(rename = "volumeUsdt")]
-    pub volume_usdt: Option<i64>,
+    pub volume_usdt: Option<f64>,
     #[serde(rename = "marketCapFD")]
-    pub market_cap_fd: Option<i64>,
+    pub market_cap_fd: Option<f64>,
     #[serde(rename = "marketCapRank")]
-    pub market_cap_rank: Option<i64>,
+    pub market_cap_rank: Option<f64>,
     #[serde(rename = "priceChange24h")]
     pub price_change24_h: Option<f64>,
 }
